Move config to new package
