+++ pkg/config/config.go
@@ -1,0 +1,6 @@
+package config
+
+type Config struct {
+	Domain  string         `yaml:"domain"`
+	Records []ConfigRecord `yaml:"records"`
+}
